chore: 🚚 Extraction des fonctions dans le dossier module
